+++ Engine/ac/game.cpp
@@ -1569,7 +1569,7 @@ void display_switch_in_resume()
 }
 
 void replace_tokens(const char*srcmes,char*destm, size_t maxlen) {
-    int indxdest=0,indxsrc=0;
+    size_t indxdest=0,indxsrc=0;
     const char*srcp;
     char *destp;
     while (srcmes[indxsrc]!=0) {
